AegisSecureRefactored/Backend/requirements.txt
fastapi
uvicorn
uvicorn[standard]
python-dotenv
motor
pydantic
pyjwt
httpx
pymongo
python-jose
passlib[bcrypt]
requests
oauthlib
google-auth
google-auth-oauthlib
google-auth-httplib2
email-validator
python-multipart
groq
uuid
datetime